% Auto-generated: do not edit by hand
\name{''DashECharts}

\alias{''DashECharts}

\title{DashECharts component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashECharts(id=NULL, bmap_token=NULL, brushSelected_data=NULL,
brush_data=NULL, click_data=NULL, clicked_bar_data=NULL,
enable_get_axisPointer_event=NULL,
enable_get_clicked_bar_data_event=NULL, event=NULL,
fun_effects=NULL, fun_keys=NULL, fun_paths=NULL,
fun_prepares=NULL, fun_values=NULL, funs=NULL,
mapbox_token=NULL, maps=NULL, n_clicks=NULL,
n_clicks_timestamp=NULL, opt_merge=NULL, option=NULL,
part_of_opt=NULL, reset_id=NULL, resize_id=NULL,
selected_data=NULL, style=NULL, yAxisResize=NULL,
yAxisRightClick=NULL, zoom_data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{bmap_token}{Character. }

\item{brushSelected_data}{Named list. }

\item{brush_data}{Named list. }

\item{click_data}{Named list. }

\item{clicked_bar_data}{Named list. }

\item{enable_get_axisPointer_event}{Logical. }

\item{enable_get_clicked_bar_data_event}{Logical. }

\item{event}{Named list. }

\item{fun_effects}{Unnamed list. }

\item{fun_keys}{Unnamed list. }

\item{fun_paths}{Named list. }

\item{fun_prepares}{Unnamed list. }

\item{fun_values}{Unnamed list. }

\item{funs}{Named list. }

\item{mapbox_token}{Character. }

\item{maps}{Named list. }

\item{n_clicks}{Numeric. }

\item{n_clicks_timestamp}{Numeric. }

\item{opt_merge}{Named list. }

\item{option}{Named list. }

\item{part_of_opt}{Named list. }

\item{reset_id}{Numeric. }

\item{resize_id}{Numeric. }

\item{selected_data}{Named list. }

\item{style}{Named list. }

\item{yAxisResize}{Named list. Receive yAxisIndex, min and max to resize yAxis}

\item{yAxisRightClick}{Named list. Send yAxisIndex, min and max after yAxis right click event}

\item{zoom_data}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

